+++ app-backend/port/src/context.rs
@@ -1,10 +1,10 @@
 pub struct RequestContext {
-    pub player_id: Option<i32>, // Adjust the type based on your identification strategy
+    pub player_id: Option<u64>, // Adjust the type based on your identification strategy
     // Other fields as needed
 }
 
 impl RequestContext {
-    pub fn new(player_id: Option<i32>) -> Self {
+    pub fn new(player_id: Option<u64>) -> Self {
         RequestContext { player_id }
     }
 }
